backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
cryptography==41.0.7
bcrypt==4.0.1
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
apscheduler==3.10.4
python-dateutil==2.8.2
yfinance==0.2.36
alembic==1.13.1
slowapi==0.1.9
